refactor: use monorepo
