--- .eslintignore
@@ -1,10 +1,0 @@
-*.sh
-node_modules
-lib
-coverage
-*.md
-*.scss
-*.woff
-*.ttf
-src/index.ts
-dist
