+++ .github/workflows/main.yml
@@ -5,8 +5,20 @@ on:
   pull_request:
     branches: [ master ]
 jobs:
-  # Lint:
-    # Lint任务
+  # This workflow contains a single job called "build"
+  Lint:
+    # The type of runner that the job will run on
+    runs-on: ubuntu-latest
+    # Steps represent a sequence of tasks that will be executed as part of the job
+    steps:
+      - uses: actions/checkout@v2
+      - uses: pnpm/action-setup@v2.1.0
+        with:
+          version: 7.2.1
+      - name: Install modules
+        run: pnpm install
+      - name: Run Lint
+        run: cd packages/smarty-ui-vite && pnpm run lint
   UnitTest:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
@@ -19,4 +31,4 @@ jobs:
       - name: Install modules
         run: pnpm install
       - name: Run Test
-        run: pnpm run test:run
+        run: cd packages/smarty-ui-vite pnpm run test:run
