+++ .github/workflows/publish.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Install modules
         run: pnpm install
       - name: Build
-        run: pnpm run build
+        run: cd packages/smarty-ui-vite && pnpm run build
       - name: "Publish to the npm registry"
         uses: primer/publish@3.0.0
         env:
@@ -28,4 +28,5 @@ jobs:
           NPM_AUTH_TOKEN: ${{ secrets.NPM_AUTH_TOKEN }} # 跟前面步骤中的 NPM_AUTH_TOKEN 保持一致
         with:
           default_branch: "publish-smarty-ui-vite"
-          dir: "."
+          dir: "packages/smarty-ui-vite"
+          # dir: "."
