+++ .husky/pre-commit
@@ -1,4 +1,6 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-pnpm run lint
+echo ====代码静态检查====
+cd packages/smarty-ui-vite
+pnpm lint
