--- demo/esm/button.html
@@ -1,26 +1,0 @@
-<head>
-  <meta charset="UTF-8">
-</head>
-
-<h1>按需加载组件</h1>
-<div id="app"></div>
-<script type="module">
-  import { createApp } from "vue/dist/vue.esm-bundler.js";
-  import {
-    SFCButton,
-    JSXButton,
-    MyButton,
-  } from "../../dist/smarty-ui.mjs";
-
-  createApp({
-    template: `
-<SButton/>
-<JSXButton/>
-<SFCButton/>
-`,
-  })
-    .component(SFCButton.name, SFCButton)
-    .component(JSXButton.name, JSXButton)
-    .component(MyButton.name, MyButton)
-    .mount("#app");
-</script>
