--- docs/.vitepress/theme/index.ts
@@ -1,9 +1,0 @@
-import Theme from 'vitepress/dist/client/theme-default/index.js'
-import SmartyUI from '../../../src/entry'
-
-export default {
-  ...Theme,
-  enhanceApp({ app }) {
-    app.use(SmartyUI)
-  },
-}
