--- docs/index.md
@@ -1,6 +1,0 @@
-## hello Vitepress
-
-```ts
-const str:string = "hello vitepress"
-
-```
