--- index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Vue3UI</title>
-  <script type="module" src="./src/index.ts" ></script>
-</head>
-<body>
-  <h1>Hello Vite And Vue</h1>
-  <div id="app"></div>
-  <!-- type="module" -->
-</body>
-</html>
