+++ package.json
@@ -1,56 +1,21 @@
 {
-  "name": "study-vue-ui",
-  "version": "1.0.4",
+  "name": "study-monorepo",
+  "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "files": [
-    "dist/"
-  ],
   "scripts": {
-    "test": "vitest",
-    "test:run": "vitest run",
-    "dev": "vite",
-    "build": "vite build",
-    "lint": "eslint --fix --ext .ts,.vue src",
-    "format": "prettier --write \"src/**/*.ts\" \"src/**/*.vue\"",
-    "docs:dev": "vitepress dev docs",
-    "docs:build": "vitepress build docs",
-    "docs:serve": "vitepress serve docs",
+    "preinstall": "node ./scripts/preinstall.js",
     "prepare": "husky install"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "dependencies": {
+    "vite": "3.0.7"
+  },
   "devDependencies": {
-    "@algolia/client-search": "^4.14.2",
     "@commitlint/cli": "17.0.2",
     "@commitlint/config-conventional": "17.0.2",
-    "@iconify-json/ic": "1.1.4",
-    "@types/node": "^18.11.9",
-    "@typescript-eslint/eslint-plugin": "^5.42.1",
-    "@typescript-eslint/parser": "^5.42.1",
-    "@vitejs/plugin-vue": "^3.2.0",
-    "@vitejs/plugin-vue-jsx": "2.0.0",
-    "@vue/eslint-config-prettier": "^7.0.0",
-    "@vue/test-utils": "2.0.2",
-    "axios": "^1.1.3",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^8.27.0",
-    "eslint-formatter-pretty": "^4.1.0",
-    "eslint-plugin-json": "^3.1.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.7.0",
-    "happy-dom": "6.0.4",
-    "husky": "^8.0.2",
-    "jsdom": "^20.0.2",
-    "prettier": "^2.7.1",
-    "terser": "5.4.0",
-    "unocss": "0.45.6",
-    "vite": "3.0.7",
-    "vitepress": "1.0.0-alpha.28",
-    "vitest": "0.25"
-  },
-  "dependencies": {
-    "vue": "3.2.37"
+    "husky": "^8.0.2"
   }
 }
