+++ packages/docs-vite/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "docs-vite",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "dev": "vite"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "study-vue-ui": "workspace:^1.0.4",
+    "vue": "3.2.37"
+  }
+}
