+++ packages/smarty-ui-vite/. npmignore
