+++ packages/smarty-ui-vite/.eslintignore
@@ -1,0 +1,10 @@
+*.sh
+node_modules
+lib
+coverage
+*.md
+*.scss
+*.woff
+*.ttf
+src/index.ts
+dist
