+++ packages/smarty-ui-vite/demo/esm/index.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+  <link rel="stylesheet" href="../../dist/style.css">
+</head>
+<body>
+  <h1>全量加载组件</h1>
+<div id="app"></div>
+<script type="module">
+  import { createApp } from "../../node_modules/vue/dist/vue.esm-bundler.js";
+  import SmartyUI from "../../dist/smarty-ui.mjs";
+
+  createApp({
+  template: `
+  <div>
+    <SButton color="blue">蓝色按钮</SButton>
+    <SButton color="green">绿色按钮</SButton>
+    <SButton color="gray">灰色按钮</SButton>
+    <SButton color="yellow">黄色按钮</SButton>
+    <SButton color="red">红色按钮</SButton>
+ </div>
+ <div style="margin-top:20px;">
+    <SButton color="blue"   icon="search" ></SButton>
+    <SButton color="green" icon="edit"></SButton>
+    <SButton color="gray"  icon="check"></SButton>
+    <SButton color="yellow"  icon="message"></SButton>
+    <SButton color="red"  icon="delete"></SButton>
+  </div>
+  `,
+})
+  .use(SmartyUI)
+  .mount('#app')
+
+  // createApp({
+  //   template: `
+  //     <SButton/>
+  //     <JSXButton/>
+  //     <SFCButton/>
+  //   `}).use(SmartyUI).mount('#app')
+
+</script>
+</body>
+</html>
+
