+++ packages/smarty-ui-vite/docs/.vitepress/theme/index.ts
@@ -1,0 +1,9 @@
+import Theme from 'vitepress/dist/client/theme-default/index.js'
+import SmartyUI from '../../../src/entry'
+
+export default {
+  ...Theme,
+  enhanceApp({ app }) {
+    app.use(SmartyUI)
+  },
+}
