+++ packages/smarty-ui-vite/docs/components/button/index.md
@@ -1,0 +1,9 @@
+# Button 按钮
+
+  <div style="margin-bottom:20px;">
+    <SButton color="blue">主要按钮</SButton>
+    <SButton color="green">绿色按钮</SButton>
+    <SButton color="gray">灰色按钮</SButton>
+    <SButton color="yellow">黄色按钮</SButton>
+    <SButton color="red">红色按钮</SButton>
+  </div>
