+++ packages/smarty-ui-vite/docs/index.md
@@ -1,0 +1,6 @@
+## hello Vitepress
+
+```ts
+const str:string = "hello vitepress"
+
+```
