+++ packages/smarty-ui-vite/publish.sh
