+++ packages/smarty-ui-vite/src/JSXButton.tsx
@@ -1,0 +1,10 @@
+import { defineComponent, h } from "vue";
+
+export default defineComponent({
+
+  name: "JSXButton",
+  render() {
+    return <button>JSX Button</button>;
+  },
+
+});
