+++ packages/smarty-ui-vite/src/SFCButton.vue
@@ -1,0 +1,10 @@
+<template>
+  <button>SFC Button</button>
+</template>
+
+<script lang="ts">
+export default {
+  name: "SFCButton",
+};
+
+</script>
