+++ packages/smarty-ui-vite/src/button/index.ts
@@ -1,0 +1,4 @@
+import Button from "./Button";
+
+// 导出Button组件
+export default Button ;
