+++ packages/smarty-ui-vite/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+      "declaration": true, /* 生成相关的 '.d.ts' 文件。 */
+      "declarationDir": "./dist/types", /* '.d.ts' 文件输出目录 */
+      "jsx": "preserve",
+      "types": ["vitest/globals"]
+  },
+  "include": [
+      "./src/**/*.*",
+      "./src/shims-vue.d.ts",
+  ],
+  "exclude": [
+      "node_modules"
+  ],
+  
+  "esModuleInterop": true,
+  "allowSyntheticDefaultImports": "true"
+}
