+++ packages/smarty-ui-vite/vite.config.ts
