+++ pnpm-workspace.yaml
@@ -1,0 +1,3 @@
+packages:
+  # all packages in subdirs of packages/ and components/
+  - 'packages/**'
