--- src/JSXButton.tsx
@@ -1,10 +1,0 @@
-import { defineComponent, h } from "vue";
-
-export default defineComponent({
-
-  name: "JSXButton",
-  render() {
-    return <button>JSX Button</button>;
-  },
-
-});
