--- src/SFCButton.vue
@@ -1,10 +1,0 @@
-<template>
-  <button>SFC Button</button>
-</template>
-
-<script lang="ts">
-export default {
-  name: "SFCButton",
-};
-
-</script>
