--- src/button/index.ts
@@ -1,4 +1,0 @@
-import Button from "./Button";
-
-// 导出Button组件
-export default Button ;
