--- src/shims-vue.d.ts
@@ -1,5 +1,0 @@
-declare module "*.vue" {
-  import { DefineComponent } from "vue";
-  const component: DefineComponent<{}, {}, any>;
-  export default component;
-}
